requirements.txt
streamlit==1.2.0
ffmpeg-python==0.2.0
requests==2.26.0
pytube==11.0.1
whisper==1.1.10
